fix(hive-activity-partner): split strings.ts (server-safe data) from useStrings.ts (client hook)
PR #118 added "use client" to strings.ts, which marked the entire module
as a Client Component boundary. Server Components that imported the
`strings` data constant (notably app/under-18/page.tsx) then received a
client reference instead of the en JSON, and the prerender failed with:

  Error occurred prerendering page "/under-18"
  TypeError: Cannot read properties of undefined (reading 'title')

Split the module so server and client paths are clean:

  - app/_lib/strings.ts     server-safe: locale JSON imports, the Strings
                            type, the strings constant, CATALOGS, and
                            pickLocale. No "use client", no React hooks.
  - app/_lib/useStrings.ts  client-only: "use client" directive, imports
                            data from ./strings, exports the useStrings
                            hook (useState + useEffect + navigator.language).

Server Components keep importing strings from ./strings unchanged.
Client Components (HiveInstallHint, HiveFirstVisitExplainer, HiveFooter,
signup/page.tsx) now import useStrings from ./useStrings.

Reverts the "use client" added to strings.ts in PR #118 in favour of
this structural split.
