+++ apps/hive-activity-partner/app/_lib/strings.ts
@@ -1,10 +1,11 @@
-"use client";
-
 // Localized strings catalog for HiveActivityPartner.
 // English is the source of truth; all other locales must mirror its shape.
 // Canonical Hive free-tier locale set: en, es, fr, ar, hi, zh, pt.
+//
+// This module is server-safe (no React hooks). Server Components import
+// `strings` directly. Client Components that need locale switching import
+// `useStrings` from "./useStrings" — that file owns the "use client" boundary.
 
-import { useEffect, useState } from "react";
 import en from "../../locales/en.json";
 import es from "../../locales/es.json";
 import fr from "../../locales/fr.json";
@@ -15,23 +16,13 @@ import pt from "../../locales/pt.json";
 
 export type Strings = typeof en;
 
-const CATALOGS: Record<string, Strings> = { en, es, fr, ar, hi, zh, pt };
+export const CATALOGS: Record<string, Strings> = { en, es, fr, ar, hi, zh, pt };
 
 export const strings: Strings = en;
 
-function pickLocale(navigatorLanguage: string | undefined): keyof typeof CATALOGS {
+export function pickLocale(navigatorLanguage: string | undefined): keyof typeof CATALOGS {
   if (!navigatorLanguage) return "en";
   const primary = navigatorLanguage.toLowerCase().split("-")[0];
   if (primary in CATALOGS) return primary as keyof typeof CATALOGS;
   return "en";
-}
-
-export function useStrings(): Strings {
-  const [s, setS] = useState<Strings>(en);
-  useEffect(() => {
-    if (typeof navigator === "undefined") return;
-    const code = pickLocale(navigator.language);
-    setS(CATALOGS[code]);
-  }, []);
-  return s;
 }
