+++ apps/hive-activity-partner/app/_lib/useStrings.ts
@@ -1,0 +1,14 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import { CATALOGS, pickLocale, strings, type Strings } from "./strings";
+
+export function useStrings(): Strings {
+  const [s, setS] = useState<Strings>(strings);
+  useEffect(() => {
+    if (typeof navigator === "undefined") return;
+    const code = pickLocale(navigator.language);
+    setS(CATALOGS[code]);
+  }, []);
+  return s;
+}
